Fix incorrect classes visibility
